requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
emoji==0.5.4
future==0.18.2
pony==0.7.13
psycopg2==2.8.4
pycparser==2.20
python-telegram-bot==12.4.2
PyYAML==5.3.1
six==1.14.0
tornado==6.0.4
validators==0.14.3
